requirements.txt
fastapi==0.104.1
starlette==0.27.0
uvicorn==0.24.0
sqlalchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
cryptography==41.0.7
jinja2==3.1.2
python-multipart==0.0.6
